+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Allow ability to add Function driers
 * Added `Number.isNumeric(input)`
 * `Array#sortByPath` will now interpret numbers as new directions
+* Added `Array#findByPath(path, value)`
 
 ## 0.3.7 (2017-07-03)